accessing firebase correctly, login doesnt work for some reason though
